Move code for frozen objects into frozen.py
