+++ polynomial/__init__.py
@@ -3,6 +3,7 @@
 (c) Yalishanda <yalishanda@abv.bg>
 """
 
+from .frozen import Freezable
 from .core import Polynomial, Monomial, Constant, ZeroPolynomial
 from .binomial import Binomial, LinearBinomial
 from .trinomial import Trinomial, QuadraticTrinomial
